+++ controllers/authController.js
@@ -3,6 +3,9 @@ const jwt = require("jsonwebtoken");
 const Joi = require("joi");
 const gravatar = require("gravatar");
 const User = require("../models/user");
+const jimp = require("jimp");
+const path = require("path");
+const fs = require("fs/promises");
 
 const registerSchema = Joi.object({
   email: Joi.string().email().required(),
@@ -59,6 +62,29 @@ async function registerUser(req, res) {
   }
 }
 
+const uploadAvatar = async (req, res) => {
+  try {
+    const { file } = req;
+    if (!file) {
+      return res.status(400).json({ message: "No file uploaded" });
+    }
+
+    const image = await jimp.read(file.path);
+    await image.cover(250, 250).write(file.path);
+
+    const newAvatarName = `${req.user._id}${path.parse(file.originalname).ext}`;
+    const newAvatarPath = path.join(__dirname, "../public/avatars", newAvatarName);
+    await fs.rename(file.path, newAvatarPath);
+
+    await User.findByIdAndUpdate(req.user._id, { avatarURL: `/avatars/${newAvatarName}` });
+
+    res.json({ avatarURL: `/avatars/${newAvatarName}` });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ message: "Server error" });
+  }
+};
+
 async function loginUser(req, res) {
   const { error } = loginSchema.validate(req.body);
   if (error) {
@@ -98,5 +124,6 @@ async function loginUser(req, res) {
 
 module.exports = {
   registerUser,
+  uploadAvatar,
   loginUser,
 };
